pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "langchain-google-vertexai>=3.2.2",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.14",
    "langsmith[claude-agent-sdk]>=0.7.32",
    "python-dotenv>=1.2.2",
]
